Make Page extendable
