+++ partials/views/page/page-hero.dust
@@ -1,5 +1,7 @@
 <section class="page__hero section is-relative pt-0 pb-0 has-background-cover" {@inlinebg id=Page.hero_image size="fullhd" /}>
-    <div class="overlay overlay--dark-50"></div>
+    {?Page.use_overlay}
+        <div class="overlay overlay--dark-50"></div>
+    {/Page.use_overlay}
 
     <div class="is-absolute has-top-50 has-right-0 has-left-0 has-transform-y--50">
         <div class="container">
